COMMON, CUDA: Adding measure time logs.

+++ cuda/src/semblance/data/CudaDeviceContext.cu
@@ -13,5 +13,5 @@ CudaDeviceContext::CudaDeviceContext(unsigned int devId) : DeviceContext(devId) 
 
 void CudaDeviceContext::activate() const {
     LOGI("Activating GPU#" << deviceId);
-    CUDA_ASSERT(cudaSetDevice(deviceId));
+    CUDA_ASSERT(cudaSetDevice(static_cast<int>(deviceId)));
 }
